Refactored - change mifos -> fineract package name
