+++ momo-backend/src/main/java/org/fineract/mobilemoney/MobileMoneyApplication.java
@@ -1,4 +1,4 @@
-package org.mifos.mobile.money;
+package org.fineract.mobilemoney;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
