+++ momo-backend/src/main/java/org/fineract/mobilemoney/api/TransactionController.java
@@ -1,12 +1,12 @@
-package org.mifos.mobile.money.api;
+package org.fineract.mobilemoney.api;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
- * Created by daniel on 2/22/17.
+ * Created by daniel on 2/3/17.
  */
 @RestController
 @RequestMapping(value="/api/v2")
-public class InitController {
+public class TransactionController {
 }
